+++ lib/src/internal/internal.dart
@@ -1,0 +1,1 @@
+export 'ticker.dart';